[FIX] web_editor: fix recognition of SVG mimetype
At some point, servers started to return `image/svg+xml; charset=utf-8`
instead of `image/svg+xml` in the `content-type` header of SVG images.
Because of this, dynamic SVGs were not properly handled anymore.

This commit adapts the mimetype comparisons to cope for this extra
`charset` info inside the mimetype.

Steps to reproduce:
- Drop an image block inside a website page.
- Replace the image.
- Search for e.g. "city".
- Pick an undraw illustration.

=> The image was lost because its `src` became `false`.

closes odoo/odoo#198961

X-original-commit: e255894abdad53691da107f79e8af10140d3b594
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>