[FIX] mail : Correct Usability Of User Settings
Issue:
------
Case:  User Settings menu is basically used to handle created users but when
clicking on the `New` button within the User Settings the form view
corresponding to that model i.e; `res.users.settings` will be opened which
is an inappropriate behavior.
Whenever a User is created that user will be reflected within that menu.
There's no meaning of having `New` button when no user is created to handle.

Solution:
-----------
Some users do not have a user settings yet, so it is a perfectly valid use case to create one.
We should not artificially prevent acceptable behavior in technical/admin menus.

Steps to reproduce:
------------------------
1. Create a database in version 16.0.
2. Go to `User Settings`.
3. Click On `New` button. Try making changes and save.
4. An Invalid Field `User` error will Occur.

This is because there's a many2One field i.e; `user_id` within the tree and form
views which is basically an existing user.
The `New` button will open a form view with no User which is a must.
And will cause the Invalid field error when trying to save any change.

Ref:
https://github.com/odoo/odoo/blob/16.0/addons/mail/views/res_users_settings_views.xml#L24

closes odoo/odoo#196844

X-original-commit: 9327589231f4bee11e61a7f101e4e703542df4f1
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>